# frontend/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Masjid Baiturrahim API",
    "version": "1.0.0",
    "description": "REST API for the mosque management app. Responses use `{ success, data?, message?, error? }` (see handlers). **Admin** routes require JWT with `admin` role. This spec is maintained for documentation; refine schemas as needed."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local backend"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Authorization: Bearer <access_token>"
      }
    }
  },
  "paths": {
...
    "/api/v1/admin/upload": {
      "post": {
        "summary": "Upload image",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete image",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },
...
  }
}